Completely done with the first version of the app

+++ myGrayCard_search.php
@@ -1,40 +1,49 @@
 <?php
     require_once 'admin_connect.php';
-    if(!isset($_POST['submit'])){
+    if(!isset($_POST['submit']))
+    {
         $query = "select * from ministryoftransport where Dealer = '$name'";
-        $result = mysqli_query($conn,$query);
-        if($result){
-            if(mysqli_num_rows($result)>0){
-                echo '
-                <thead>
-                    <tr>
-                    <th scope="col">ID</th>
-                    <th scope="col">Owner</th>
-                    <th scope="col">R\N</th>
-                    <th scope="col">Action</th>
-                    </tr>
-                </thead>';
+        if($query != null)
+        {
+            $result = mysqli_query($conn,$query);
+            if($result){
+                if(mysqli_num_rows($result)>0){
+                    echo '
+                    <thead>
+                        <tr>
+                        <th scope="col">ID</th>
+                        <th scope="col">Owner</th>
+                        <th scope="col">R\N</th>
+                        <th scope="col">Action</th>
+                        </tr>
+                    </thead>';
 
-            while($record=mysqli_fetch_assoc($result)){
-                $id = $record['id'];
-                $Owner = $record['Owner'];
-                $carReg = $record['carReg'];
-                echo '
-                    <tr>
-                        <th scope="row">'.$id.'</th>
-                        <td>'.$Owner.'</td>
-                        <td>'.$carReg.'</td>
-                        <td>
-							<button class="btn-success"><a href="myGrayCard_Show.php?showid='.$id.'" class="text-light">Show</a></button>
-							<button class="btn-primary"><a href="myGrayCard_update.php?updateid='.$id.'">Update</a></button>
-							<button class="btn-danger"><a href="myGrayCard_delete.php?deleteid='.$id.'" class="text-light">Delete</a></button>
-						</td>
-                    </tr>
-                   ';
+                    while($record=mysqli_fetch_assoc($result))
+                    {
+                        $id = $record['id'];
+                        $Owner = $record['Owner'];
+                        $carReg = $record['carReg'];
+                        echo '
+                            <tr>
+                                <th scope="row">'.$id.'</th>
+                                <td>'.$Owner.'</td>
+                                <td>'.$carReg.'</td>
+                                <td>
+                                    <button class="btn-success"><a href="myGrayCard_Show.php?showid='.$id.'" class="text-light">Details</a></button>
+                                    <button class="btn-primary"><a href="myGrayCard_update.php?updateid='.$id.'">Update</a></button>
+                                    <button class="btn-danger"><a href="myGrayCard_delete.php?deleteid='.$id.'" class="text-light">Delete</a></button>
+                                </td>
+                            </tr>
+                        ';
+                    }
+                }
+            }
+            else{
+                echo "NO DATA FOUND!";
             }
         }
+
     }
-}
 
     if(isset($_POST['submit'])){
         $search = $_POST['searchRequest'];
@@ -63,7 +72,7 @@
                             <td>'.$Owner.'</td>
                             <td>'.$carReg.'</td>
                             <td>
-								<button class="btn-success"><a href="myGrayCard_Show.php?showid='.$id.'" class="text-light">Show</a></button>
+								<button class="btn-success"><a href="myGrayCard_Show.php?showid='.$id.'" class="text-light">Details</a></button>
 								<button class="btn-primary"><a href="myGrayCard_update.php?updateid='.$id.'">Update</a></button>
 								<button class="btn-danger"><a href="myGrayCard_delete.php?deleteid='.$id.'" class="text-light">Delete</a></button>
 							</td>
